improv: move load_dictionary tests into a separated file
